works! The countviolations still isn't working but the package is fixing the semantic line breaks
Need more test cases on tables and other oddities
